+++ fwdpp/types/Makefile.am
@@ -1,4 +1,5 @@
 pkgincludedir=$(prefix)/include/fwdpp/types
 
-pkginclude_HEADERS=mutation.hpp
+pkginclude_HEADERS=mutation.hpp \
+				   nested_forward_lists.hpp
